Need to truncate gce serviceaccounts to max 30 characters
