+++ pkg/truncate/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
@@ -6,4 +6,10 @@ go_library(
     importpath = "k8s.io/kops/pkg/truncate",
     visibility = ["//visibility:public"],
     deps = ["//vendor/k8s.io/klog/v2:go_default_library"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["truncate_test.go"],
+    embed = [":go_default_library"],
 )
